refactor(restore): centralice the logic in the same place
